requirements.txt
python-dotenv
pinecone
matplotlib